Fix example, move propagation_test
